pyproject.toml
[tool.poetry]
name = "netguru-movies"
version = "0.1.0"
description = "Recruitment task"
authors = ["kuter <contact@devktr.pl>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.0"
djangorestframework = "^3.11.1"
markdown = "^3.2.2"
django-filter = "^2.3.0"
django-appconf = "^1.0.4"
requests = "^2.24.0"
drf-writable-nested = "^0.6.1"
drf-yasg = "^1.17.1"
psycopg2-binary = ">=2.8"

[tool.poetry.dev-dependencies]
pre-commit = "^2.6.0"
wemake-python-styleguide = "^0.14.1"
factory-boy = "^3.0.1"
coverage = "^5.2.1"
responses = "^0.10.16"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
